Update to version 4

+++ zscores/code/submit.py
@@ -1,32 +1,33 @@
 import pandas as pd
 import subprocess
+import time
+import os
 from jinja2 import Template
 
 submission = Template("""#!/bin/sh
-#SBATCH --job-name={{batch}}
-#SBATCH --nodes=1
-#SBATCH --cpus-per-task=2
-#SBATCH --partition=sbinlab_ib2
-#SBATCH --mem=2GB
-#SBATCH -t 100:00:00
-#SBATCH -o {{batch}}_out
-#SBATCH -e {{batch}}_err
+#PBS -W group_list=ku_10001 -A ku_10001
+#PBS -N {{batch}}
+#PBS -m n
+#PBS -l nodes=1:ppn=1:thinnode
+#PBS -l mem=1gb
+#PBS -l walltime=72:00:00
+#PBS -e {{batch}}_err
+#PBS -o {{batch}}_out
 
-source /groups/sbinlab/giulio/.bashrc
+source /home/people/giutes/.bashrc
+conda activate idrome
 
-conda activate nardini
+cd $PBS_O_WORKDIR
 
-echo $SLURM_CPUS_PER_TASK
-
-echo $SLURM_CPUS_ON_NODE
-
-python ./analyse.py --batch {{batch}}
+python ./calc_seq_prop.py --batch {{batch}}
 """)
-
 
 df_idrome = pd.read_csv('IDRome_DB.csv',index_col=0)
 
-for batch in range(0,df_idrome.shape[0],50):
-    with open(f'{batch:d}.sh', 'w') as submit:
-        submit.write(submission.render(batch=batch))
-    subprocess.run(['sbatch',f'{batch:d}.sh'])
+for batch in range(0,df_idrome.shape[0],20):
+    if not os.path.isfile(f'seq_prop_{batch:d}.csv.gz'):
+        print(batch)
+        with open(f'{batch:d}.sh', 'w') as submit:
+            submit.write(submission.render(batch=batch))
+        subprocess.run(['qsub',f'{batch:d}.sh'])
+    time.sleep(.6)
